backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
pydantic
python-dotenv
asyncpg
